Combine two APIs (request and retrieve) into retrieve
The separation of the two APIs is an optimization, considering that an
object saved into some format might be accessed multiple times, so that
it needs not to be saved multiple times. However, accessing the same
file multiple time might be a rare case, and we might figure out a way
to avoid using tmp files at all in the future.