Add plugins dir to ImageJ search path
Prior to this commit, ImageJ would not load any plugins in the plugins
directory.  This patch adds the plugin directory to ImageJ's search,
so it properly loads the plugins for the scripting environment.

Note: Test to follow once I can figure out a good way of measuring if
this works or not.