Add ability to spin up ImageJ without local copy
The scyjava library lets us add more endpoints to the classpath,
retrieved from remote Maven repositories, before spinning up Java.
In this way, we can download ImageJ on demand.

If a version is given, add endpoint net.imagej:imagej:<version>.
If an endpoint is given, add that endpoint before creating the gateway.

The unit tests are also updated to support testing ImageJ from Maven
when a local installation directory is not provided.

Closes #9.