Incorporate imagej1 windows
This change makes it easy to work with ImageJ1 windows without needing to reference jnius or outside toolboxes.

First, it sets a new module ij.window to be the WindowManager.  This saves the awkward step of classifying the WindowManager outside the ij.init declaration.

Second, it creates a function to convert a numpy array to a window in ij.py.  The UIService is capable of doing this already, but it is not an intuitive way of transitioning between numpy and imagej.  This makes it much simpler and more clear how you are supposed to do so.