Accept lists of endpoints and strip out whitespace
This commit lets a user enter in a list of Maven endpoints, rather than
a single string.  This will make it more pythonic to enter in multiple
endpoints.

It also strips whitespace from a string that is a Maven endpoint.  This
is in case someone enters e.g.
'net.imagej:imagej + net.imagej:imagej-legacy', we assume they meant the
endpoint 'net.imagej:imagej+net.imagej:imagej-legacy'