Fix imagej/fiji version selection
This commit moves the import of imglyb after setting the
endpoints. If 'net.imglib2:imglib2-limglyb:1.0.0' is the first
endpoint in the list, imagej version 2.1.0 is loaded regardless of
the specified version desired. This can be resolved by re-ordering
the sort of endpoints in scyjava/__init__.py or, in this case, moving
the import of imglyb after the other endpoints have been added.