Split imagej.init into two parts
* _create_jvm sets up and starts the JVM.
* _create_gateway makes an instance of net.imagej.ImageJ.

It is sometimes useful to be able to do the former without the latter.
In particular, on macOS, the _create_jvm with headless=False does not do
anything AWT-specific, leaving us in a good state to subsequently call
jpype.setupGuiEnvironment to display the ImageJ GUI without hanging.

Example code:

  import imagej, jpype
  imagej._create_jvm(ij_dir_or_version_or_endpoint=None, headless=False, add_legacy=True)
  jpype.setupGuiEnvironment(lambda: imagej._create_gateway().ui().showUI())

This code requires you have pyobjc installed via pip
(the various conda packages did not work for me).

We can still do better, but this is a good step forward.