Make imagej.init accept one of three modes
Instead of a headless flag, we now have three modes: HEADLESS, GUI, and
INTERACTIVE. And we behave consistently across all platforms as much as
we reasonably can. The main exception is that interactive mode is not
available on macOS, due to limitations in Apple's app threading model.

Of course, it's possible that these supposed limitations are not real,
and all we need to do is call AppHelper.stopEventLoop() at the right
time or something, but I am clueless about macOS programming, so
for now we just apologize and tell people the mode is unavailable.

Closes #23. At least: it's good enough for now.