Catch 'C' and 'T' dimensions
If 'C' and 'T' or their lowercase counterparts are passed to 'net.imagej.axis.Axes'
to obtain the axis type, 'C' and 'T' returns their own axis type and not Axis.CHANNEL
or Axis.TIME as they should be. These get lost in the organization step as we check for
Channel and Time axes. Inadvently, the 'C' and 'T' axes are treated as
extra beyond the magic 5 and are inserted after the 'Z' dimension.