Start fixing the plain ImageJ2 mode
We cannot have hard imports of original ImageJ classes.
As soon as add_legacy=False, such imports break everything.

It still doesn't fully work, though, due to discrepancies across the
codebase in the pre- and post-conditions of the ij.legacy property
behavior. Should it raise, or not raise, when LegacyService is missing?