Remove dtype function from ImagePlus for now
This partially reverts commit 62178c192c5c9ed9a6a8172fb27c3ca2e3a58d78.

As per this discussion:
  https://gitter.im/imagej/pyimagej?at=6269dc98949ae940067e9830

@gselzer pointed out that slicing an ImagePlus down to a single element
should -- in the same way that numpy and ImgLib2 images behave -- return
the element as an instance of the same class that dtype reports. For
ImageJ, this is not so easy, though, because ImageJ does not have
sample value container types like ImgLib2 does. So we would need to:

A) Invent some ImageJ-specific ones; or
B) Reuse the numpy ones; or
C) Use plain Python types but that is ambiguous (GRAY8 vs GRAY16); or
D) Break the convention.

See also #194.