Fix the implementation of run_plugin
And add a test for it that runs with plain ImageJ2+ImageJ (as opposed
to test_plugins_load_using_pairwise_stitching, which requires Fiji).

This function was actually calling run_macro (!!).
It is much better to use IJ.run! So that's what it does now.

The test_plugins_load_using_pairwise_stitching implementation has also
been updated to avoid using macros, in favor of direct ImageJ API.

Closes #126.