Use "ImageJ2 developers" as copyright holders
This is what we are doing across the SciJava codebase.
It is much easier to keep aligned. People can ask Git to learn
who contributed what to what files. (On the Java side, we also
use author tags per file in the javadoc; maybe we should do
the equivalent of that in the Python source files as well?)

For those who claim that a blanket statement like this isn't legally
appropriate: perhaps not, but neither is listing original authors.
Legally, you must list the actual entities with the copyright, which in
the case of work-for-hire typically means the companies, universities,
etc., employing the people doing the work, rather than those
people themselves. The ImageJ2 project used to do this ("Copyright 20xx
Board of Regents of the University of Wisconsin-Madison, Max Plank
Institute for Cell Biology and Genetics, ...") but it was still only
maintained that way per repository, not per file, because maintaining it
manually per file is much less feasible to automate easily, even though
it is more legally correct to do so.

Of course, I am not a lawyer, but in my opinion, requiring everyone who
contributes to a project to: A) figure out which entity owns the
copyright for their work before their contribution can start being
accepted, and then B) maintain a distinct copyright header block for
every single file in the repository according to which files that
person/entity touched, is a terrible extra burden with a chilling effect
on contributions, and if you look across the landscape of publicly
available open-source software, almost no one does this. And for those
that do, I claim there are likely many errors across their codebase.
So let's not waste any more of our time with this nonsense.