Fix invalid path error if path contains "~"
If the path to the local ImageJ/Fiji installation
contains the very convenient "~" short for "/home/user"
the JVM will fail to initialize. This is because
the path fails the os.path.isdir() check. Using
os.path.expanduser() resolves this issue (and does not
effect paths that do not contain "~").