Modularize the tests more cleanly
There was a lot of stuff lumped into test_imagej.py. So many fixtures!

Furthermore, we were grouping test functions into classes, when
it works just as well to keep test functions at the global level.
There are reasons to group test functions into classes (see
https://stackoverflow.com/a/50028551/1207769 for a couple of them),
but for our usage, we do not need it.

Now, every file has (up to) four sections in consistent order:

1. Imports
2. Fixtures
3. Helpers
4. Tests

This order is consistent with the pytest documentation,
which gives examples that show fixtures followed by tests.

Common fixtures still live in the toplevel conftest.py,
which is the canonical pytest best practice. There are more
complex conftest.py organizations across multiple files (see
https://stackoverflow.com/a/34520971/1207769 for details)
but it would be overkill for this project).