Reduce the size of the test ImgPlus
We don't need to use all prime numbers for the dimensions; it's probably
enough that each dimensional length is distinct, so that we can tell
which dimension is which by its unique length, in case of bugs.

This cuts pytest execution time by half on my system: ~10s to ~5s.