Refactor dims._assign_axes()
This commit refactors dims._assign_axes() to use
imagej specific metadata attached to a given
xarray.DataArray. If the "imagej" attribute is present
in the xarray's global attributes then information
like scale, origin and the type of CalibratedAxis
(e.g. DefaultLinearAxis, EnumeratedAxis, etc...) can be used
to assign the correct calibrated axis per given dimension
during the net.imagej.Dataset conversion process. If the
desired calibrated axis is not available or the axis is
unknown then we fall back to a DefaultLinearAxis and attempt
to get scale/origin data from the coordinates.