Add _update method to MetadataAccessor
The _update() method runs any time the MetadataAccerssor is accessed.
This allows us to update the metadata base by checking the state
of the backing xarray.DataArray. So for example if dimensions change
order or are dropped the "scifio.metadata.image" (if present) metadata
should reflect these changes where appropriate.