Remove Java to Python stdout and stderr forwarding
We noticed that running macros and scripts (e.g. ij.py.run_script())
produces a duplicate stdout and stderr. We traced the duplicating
output to the OutputListener forwarding removed in this commit. But
why were we forwarding the stdout and stderr in the first place? Looking
at commit db9bae3c we can see that we carried the OutputListener
forwarding from pyjnius to JPype. JPype already handles stdout and
stderr forwarding, thus a duplicate is generated. Apparently we
never noticed this. This commit removes the duplicate forwarding.