add conda-forge as channel to env create command
In my case, I did not want to install Miniforge3. I did not understand why it's necessary as I already have conda and mamba installed.

When trying your mamba create command, pyimagej could not be found.
A simple `-c conda-forge` added to the create command solves the problem without the need for Miniforge3, which requires admin priveleges to install.