Add check if running on main thread for macOS
Jaunch now enables a true interactive experince
for macOS users, but only if PyImageJ is launch in
a seperate thread (i.e. not the main thread). This commit
adds an additional check for which thread we are currently on.

If on the main thread, fail like before and prevent interactive
mode on macOS. Otherwise, proceed as usual.