Add check to determine if current thread is main
This commit introduces a check to determine if the current
running thread is the main thread. If True, then we continue to block
interactive mode on macOS (it is not possible to share the main thread
with GUI event loop needed for the ImageJ GUI). If False, then the
current thread is *not* the main thread (e.g. a jaunched session) and
interactive mode for macOS can proceed.