Fix xarray dimension renaming
xarray used to produce a UserWarning when renaming DataArrays with
coordinates. At the time this warning would print to screen anytime
ij.py.to_xarray() was used with a dimension order sequence. In order to
address this issue I followed the warning's suggestion and used
swap_dims and set_index to by pass this issue. Unfortunately I did not
see [this discussion](https://github.com/pydata/xarray/discussions/7950)
indicating the proper remedy of supressing the warning. This warning was
slated for removal as, I presume, the xarray team decided to maintain
its behavior.

As such *my* fix is bad and produces this error when trying to rename
any axis that has a coordinate:

ValueError: replacement dimension 'row' is not a 1D variable along the
old dimension 'ch'

This commit fixes this bug by simply reverting back to our old behavior
of using the rename() method of xarray.DataArray.

Closes #320