Simplify Named RAI name conversion
When you use ij.py.from_java, you open the door to unwanted conversion
behavior. As an example, if jobj is linked to an ImagePlus, the
conversion of the name can utilize imagej-legacy's
StringToImagePlusConverter, making the resulting xarray.name an ndarray.

This was the source of https://github.com/imagej/napari-imagej/issues/304