Fix Java fetch with cjdk when no Java is present
This commit fixes a bug in _guess_java_version() on systems
where *no* Java is present on the path. In this scenario we
attempt to get the JVM version, which asks jpype for the JVM
path as part of the request. Because no JVM is present or running
a `JVMNotFoundException` is thrown by jpype (not a RuntimeException).

This commit catches the right exception and also
initializes the version_digits variable to None (preventing an
undefined variable error if the try except fails).

I tested this on a fresh VM of Ubuntu server that has no system wide
Java installation.