_guess_java_version: Use _logger.debug instead
_log_exception's signature suggests that it is only meant to handle Java
exceptions, however in many places it is called with Python exceptions.
This has the unintended effect of starting up the JVM whenever debug
logging is enabled.

We should thus be careful to either (a) use _log_exception everywhere
but add case logic to avoid starting up the JVM early, or (b) use
_log_exception ONLY when providing a Java exception.