Fix run_macro and run_script return types
This commit fixes the return type of run_macro and run_script
to return the promised Java map, instead of a ScriptModule
instance. The doc strings for both functions has been clarified
to describe the expected outputs. I also noticed that we were
not testing these functions, so I added a simple tests for both
run_macro and run_script.

Closes #316