Implement ScriptModule functions for Java HashMap
With commit b1965c6 I fixed the outputs of `run_script`
and `run_macro` by returning the promised Java HashMap
instead of an instance of a `ScriptModule`. This change
however (1) changes the API behavior and (2) breaks any
script that depends on the `getOutput()` and `getOutputs`
calls. To alleviate this headache I made, we agreed to
implement these methods on the Java HashMap, thereby
not breaking scripts calling for outputs. However the
`getInput` and `getInputs` functions are not implemented
as they would exisit in a different Hash map. Calling
these will raise a `NotImplementedError`.

This commit also adds the appropriate tests.