Update pyproject.toml for uv and RTD
For uv we should use dependency-groups. This patch updates our scripts
to use the appropriate dependency-groups when running.

It also adds back project.optional-dependencies, which is only used for
ReadTheDocs due to its lack of native support for uv at the time.

Development directions are updated as we do not actually need to
manually build a uv virtual environment, now that the appropriate
commands are using dependency-groups.