Make PyImageJ work without the labeling dependency
The labeling package depends on scipy, which is a big dependency,
which does not yet have binary wheels for Python 3.14.

We don't want to mark labeling as optional; i.e., we want it to be
installed by default. But if it's somehow not actually present -- e.g.
if you want to try PyImageJ with Python 3.14 by commenting out
the labeling dependency from pyproject.toml -- this patch disables
labeling conversions surgically when the dependency is missing.