Check if the JVM is running before checking version
This commit fixes a bug where we fail to reach the
cjdk fetch code if no JVM is installed first. Specifically,
when `_prepare_to_unlock_modules()` is called it first
tries to collect the major version calling `_guess_java_version()`.
Once here, we hit the try/except and then return on the exception,
instead of hitting the JDK fetch code. All we needed
to do was check if there is a running JVM first. If there
is not, we bypass the try/except and hit cjdk.

Closes #337